requirements.txt
aiogram~=3.6.0
aiogram_dialog~=2.1
pydantic~=2.7.1
tortoise-orm~=0.21.6